Almost working DFA?
Appears to train, I'll post distributions demonstrating this in the slack soon-ish. Cleaned up the code a bit before uploading so I didn't keep the data from the run. If there's a problem keeping this from working then it's most likely in how the gradients for the last layer/its biases are calculated. I did that exactly as how gradients are normally calculated for every variable, but the gradients at step 0 are still all zero for some reason, even after switching back to a normal initial distribution for that layer(vs zero for every other and in previous attempts).