Trains a bit
Probably going to have to call this one a dead end. I found the problems I was having before but this still doesn't really work. It trains well initially, but loss seems to hit an asymptote or something after a few thousand steps. The initial improvements could literally just be the last layer training. I think the problem is probably how I've been getting dActivation/dWeights. Still uploading the code in case anyone wants to take a look at it or test it themselves.